+++ examples/resources/kilhog_network/resource.tf
@@ -2,10 +2,7 @@ resource "kilhog_network" "production" {
   name        = "production"
   description = "Production network"
 
-  tags = [
-    {
-      key   = "environment"
-      value = "production"
-    },
-  ]
+  tags = {
+    environment = "production"
+  }
 }